pyproject.toml
[tool.poetry]
name = "python-demo"
version = "0.1.0"
description = ""
authors = ["Dương Mạnh Hoàng <manhhoang3151996@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
Pillow = "^8.1.0"
pytesseract = "^0.3.7"
numpy = "^1.20.1"
opencv-python = "^4.5.1"
pdf2image = "^1.14.0"
yapf = "^0.30.0"
isort = "^5.7.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
